+++ modules/public_pages/erp/order/sales_order/models/sordersSearch.php
@@ -95,7 +95,7 @@ class sordersSearch extends BaseSearch
 		$search->addSearchField(
 			'status',
 			'status',
-			'select',
+			'multi_select',
 			'',
 			'advanced'
 			);